having trouble getting associations right
billturner, save me from myself

+++ db/migrate/20100922003035_create_tickets.rb
@@ -2,13 +2,13 @@ class CreateTickets < ActiveRecord::Migration
   def self.up
     create_table :tickets do |t|
       t.references :event
-      t.int :price
-      t.int :available
+      t.float :price
+      t.integer :available
       t.datetime :date_open
       t.datetime :date_closed
       t.boolean :package
       t.references :generates_ticket
-      t.int :generates_number
+      t.integer :generates_number
 
       t.timestamps
     end
